Upload
This project uses Tableau to analyze data from the Indian Premier League (IPL). The data includes information on player performance, team performance, and match results. The project uses a variety of visualizations to explore the data and identify trends. The results of the analysis can be used to improve team performance and make better decisions about player selection.

The project was created by me and is available for download on GitHub. The project includes a Tableau workbook that contains all of the visualizations created for the project. The workbook also includes a README file that provides more information about the project.

The project was created using the following steps:

The data was collected from the IPL website.
The data was cleaned and prepared for analysis.
The data was analyzed using Tableau.
The visualizations were created in Tableau.
The workbook was created and published on GitHub.
The project has the following benefits:

The project provides insights into player performance, team performance, and match results.
The project can be used to improve team performance and make better decisions about player selection.
The project is open-source and can be used by anyone.
The project has the following limitations:

The project is based on historical data and may not be accurate for future predictions.
The project is limited by the quality of the data that was collected.
The project is not intended to be used for commercial purposes.

Overall, this project is a valuable resource for anyone interested in learning more about the Indian Premier League. The project provides insights into player performance, team performance, and match results. The project can be used to improve team performance and make better decisions about player selection. The project is open source and can be used by anyone.